memory-system/requirements.txt
mcp[cli]
pinecone
openai
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
psutil
schedule
sentence-transformers
